format the code and add the cmakelist.txt
